enterprise-apps/finance-accounting/receipt-ocr/requirements.txt
streamlit==1.28.0
pandas==2.1.1
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
